requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.62.0
transformers>=4.30.0  # For pretrained CLIP text model
# FAISS for vector similarity search
# Choose one:
# faiss-cpu>=1.7.3  # For CPU-only systems
#faiss-gpu>=1.7.3   # For GPU acceleration (requires CUDA)
